+++ example/lib/main.resource.dart
@@ -1,0 +1,18 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+// **************************************************************************
+// ResourceGenerator
+// **************************************************************************
+
+// **************************************************************************
+// if you add new resource file, recommend to run clean first：
+// flutter packages pub run build_runner clean
+//
+// run following command to generate resource class：
+// flutter packages pub run build_runner build --delete-conflicting-outputs
+// **************************************************************************
+
+abstract class R_Image {
+  static const String CLOSE_A = 'assets/image/close_a.png';
+  static const String CLOSE_B = 'assets/image/close_b.png';
+}