+++ lib/src/file_handler.dart
@@ -2,21 +2,21 @@ import 'dart:io';
 
 import 'package:flutter_resource_generator/resource_config.dart';
 import 'package:flutter_resource_generator/src/util.dart';
+import 'package:path/path.dart';
 
 class FileHandler {
-
   final ResourceConfig _config;
-  final List<String> _imageFiles = [];
-  final List<String> _fontFiles = [];
-  final Map<String, List<String>> _otherFiles = {};
+  final Set<String> _imageFiles = Set();
+  final Set<String> _fontFiles = Set();
+  final Map<String, Set<String>> _otherFiles = {};
 
   FileHandler(this._config);
 
-  List<String> get imageFiles => _imageFiles;
+  Iterable<String> get imageFiles => _imageFiles;
 
-  List<String> get fontFiles => _fontFiles;
+  Iterable<String> get fontFiles => _fontFiles;
 
-  Map<String, List<String>> get otherFiles => _otherFiles;
+  Map<String, Iterable<String>> get otherFiles => _otherFiles;
 
   void parseFiles() {
     Directory root = Directory(_config.resourcePath);
@@ -28,8 +28,12 @@ class FileHandler {
           var name = getNameWithoutExtensionFromPath(path);
           if (extension != null && name != null) {
             if (!(_config.ignoreExtensions?.contains(extension) ?? false)) {
-              if (ResourceConfig.IMAGE_EXTENSIONS.contains(extension) || (_config.extraImageExtensions?.contains(extension) ?? false)) {
-                path = path.replaceAll(RegExp(r'[1-9]\.0x/'), '');
+              if (ResourceConfig.IMAGE_EXTENSIONS.contains(extension) ||
+                  (_config.extraImageExtensions?.contains(extension) ??
+                      false)) {
+                path = path.replaceAll(
+                    RegExp(r'[1-9]\.0x' + (separator == '/' ? '/' : r'\\')),
+                    '');
                 _imageFiles.add(path);
               } else if (extension == '.ttf') {
                 if (_config.handleFontFile) {
@@ -38,7 +42,7 @@ class FileHandler {
               } else {
                 var filePathList = _otherFiles[extension];
                 if (null == filePathList) {
-                  filePathList = [];
+                  filePathList = Set();
                   _otherFiles[extension] = filePathList;
                 }
                 filePathList.add(path);
@@ -49,5 +53,4 @@ class FileHandler {
       }
     }
   }
-
 }