+++ test/flutter_resource_generator_test.dart
@@ -1,9 +1,1 @@
-import 'package:flutter_test/flutter_test.dart';
-
-import 'package:flutter_resource_generator/flutter_resource_generator.dart';
-
-void main() {
-  test('adds one to input values', () {
-
-  });
-}
+void main() {}